Work so far
I’m missing the Json file?? I forgot where that comes from. I thought
it was when I referenced the Jquery library….

Also—I can’t get my class myButton to work in the style sheet, but I
need to move on. (Note that it isn’t displaying in orange like it
should).

Also also—for the API part I was thinking of having a sound element
play when you press the button (maybe instead of the alert…or both if I
can figure that out). Does that count as an API?